create a rental folder in dto, refactor the code
