+++ api/src/main/java/com/chatop/api/dto/rental/RentalCreateDto.java
@@ -1,4 +1,4 @@
-package com.chatop.api.dto;
+package com.chatop.api.dto.rental;
 
 import java.math.BigDecimal;
